At the point where I can now strip existing bazel files and autogen intefaces.

+++ tools/workspace_patch.py
@@ -43,18 +43,17 @@ def main():
         "--workspace",
         required=True,
         type=str,
-        help="Bazel 'ros' module release version, e.g. rolling.2026-01-21",
-    )
-    parser.add_argument(
-        "--packages",
-        nargs='+',
-        type=str,
-        help="List of packages to include in the workspace",
+        help="Bazel 'ros' module release version, e.g. rolling.2026-01-21.rcr.1",
     )
     parser.add_argument(
         "--overwrite",
         action="store_true",
         help="Overwrite the workspace directory if it exists",
+    )
+    parser.add_argument(
+        "--dry-run",
+        action="store_true",
+        help="Don't actually create the new patch release, just print what we would do",
     )
     args = parser.parse_args()
 
@@ -64,26 +63,26 @@ def main():
         # NEW: rolling.2026-01-21.bcr.3   (what we are adding)
         modules_dir = args.working_directory / "modules"
 
-        # Find the release from the given workspace
+        # Find the release from the given workspace.
         ref_patch = find_release(args.workspace)
         ref_dir = args.working_directory / "workspace" / ref_patch
-        sp.write("> Found REF: {0}".format(ref_patch))
+        sp.write("> Found REF: {0} (this is the reference bootstrapped release)".format(ref_patch))
 
         # Given the release, we need to find the OLD and NEW patch versions.
         old_patch = find_latest_patch(args.working_directory, ref_patch)
-        sp.write("> Found OLD: {0}".format(old_patch))
-
+        sp.write("> Found OLD: {0} (this is the last patch release)".format(old_patch))
+        
         # Make sure the workspace exists
         new_dir = args.working_directory / "workspace" / args.workspace
         if not new_dir.exists():
             raise RuntimeError("Workspace {0} does not exist".format(args.workspace))
-        sp.write("> Found NEW: {0}".format(args.workspace))
+        sp.write("> Found NEW: {0} (this is what we're about to release)".format(args.workspace))
 
         # Get all the old packages
         old_module_file = args.working_directory / "modules" / "ros" / old_patch / 'MODULE.bazel'
         old_packages = scan_module_for_dependencies(old_module_file, modules_dir)
         old_packages["ros"] = old_patch
-        sp.write("> Found {0} OLD packages".format(len(old_packages)))
+        sp.write("> Found {0} packages in the OLD patch release".format(len(old_packages))) 
 
         # Find all the packages in the REF release
         # sp.write("> Scanning REF release for packages")
@@ -96,15 +95,11 @@ def main():
             d.name.removesuffix("+") : d.name for d in (new_dir / "vendor").iterdir()
             if d.is_dir() and d.name.endswith("+")
         }
-        sp.write("> Found {0} NEW packages".format(len(packages)))
-        if args.packages is not None:
-            sp.write("> Including packages: {0}".format(", ".join(args.packages)))
-            packages = {key: value for key, value in packages.items() if key in args.packages}
-            sp.write("> Filtered to {0} NEW package".format(len(packages)))
+        sp.write("> Found {0} packages in NEW patch release".format(len(packages)))
 
         # Calculating diffs for all packages
         updated_modules = {}
-        sp.write("> Calculating which modules need updating")
+        sp.write("> Calculating differences between OLD and NEW")
         for package_name, package_dir in sorted(packages.items()):
             #sp.write("> Processing {0}".format(package_name))
 
@@ -112,13 +107,11 @@ def main():
             ref_package_dir = ref_dir / "vendor" / package_dir
             new_package_dir = new_dir / "vendor" / package_dir
             new_patches, new_overlays = get_module_diff(ref_package_dir, new_package_dir)            
-            #sp.write("  + Found {0} patches and {1} overlays in NEW".format(len(new_patches), len(new_overlays)))
 
             # Load the patches and overlays from the OLD module directory.
             old_package_version = old_packages[package_name]
             old_patches, old_overlays = load_patches_and_overlays(
                 args.working_directory / "modules" / package_name / old_package_version)
-            #sp.write("  + Found {0} patches and {1} overlays in OLD".format(len(old_patches), len(old_overlays)))
 
             # If there are no changes, leave the module as-is.
             if (old_patches == new_patches) and (old_overlays == new_overlays):
@@ -128,35 +121,36 @@ def main():
             new_package_version = increment_version(old_package_version)
             sp.write("  + Incrementing {0} version from {1} to {2}".format(
                 package_name, old_package_version, new_package_version))            
-            
-            # Copy the old module
-            old_module_dir = args.working_directory / "modules" / package_name / old_package_version
-            new_module_dir = args.working_directory / "modules" / package_name / new_package_version
-            shutil.copytree(old_module_dir, new_module_dir, dirs_exist_ok=True)
-
-            # Replace the overlay and patches files.
-            shutil.rmtree(new_module_dir / "patches", ignore_errors=True)
-            shutil.rmtree(new_module_dir / "overlay", ignore_errors=True)
-            for patch_name, patch_content in new_patches.items():
-                patch_path = new_module_dir / "patches" / patch_name
-                patch_path.parent.mkdir(parents=True, exist_ok=True)
-                with open(patch_path, 'w') as f:
-                    f.write(patch_content)  
-            for overlay_name, overlay_content in new_overlays.items():
-                overlay_path = new_module_dir / "overlay" / overlay_name
-                overlay_path.parent.mkdir(parents=True, exist_ok=True)
-                with open(overlay_path, 'w') as f:
-                    f.write(overlay_content)  
-
-            # Update the source.json file.
-            regenerate_integrity_hashes(new_module_dir)
-
-            # Update the metadata.json file
-            metadata_json_path = args.working_directory / "modules" / package_name / "metadata.json"
-            add_version_to_metadata_json(metadata_json_path, new_package_version)
-
-            # List this as an updated module.
             updated_modules[package_name] = new_package_version
+
+            # Only touch file if we are not in dry-run mode.
+            if not args.dry_run:
+
+                # Copy the old module
+                old_module_dir = args.working_directory / "modules" / package_name / old_package_version
+                new_module_dir = args.working_directory / "modules" / package_name / new_package_version
+                shutil.copytree(old_module_dir, new_module_dir, dirs_exist_ok=True)
+
+                # Replace the overlay and patches files.
+                shutil.rmtree(new_module_dir / "patches", ignore_errors=True)
+                shutil.rmtree(new_module_dir / "overlay", ignore_errors=True)
+                for patch_name, patch_content in new_patches.items():
+                    patch_path = new_module_dir / "patches" / patch_name
+                    patch_path.parent.mkdir(parents=True, exist_ok=True)
+                    with open(patch_path, 'w') as f:
+                        f.write(patch_content)  
+                for overlay_name, overlay_content in new_overlays.items():
+                    overlay_path = new_module_dir / "overlay" / overlay_name
+                    overlay_path.parent.mkdir(parents=True, exist_ok=True)
+                    with open(overlay_path, 'w') as f:
+                        f.write(overlay_content)  
+
+                # Update the source.json file.
+                regenerate_integrity_hashes(new_module_dir)
+
+                # Update the metadata.json file
+                metadata_json_path = args.working_directory / "modules" / package_name / "metadata.json"
+                add_version_to_metadata_json(metadata_json_path, new_package_version)
         sp.write("> Need to update {0} modules with new content".format(len(updated_modules)))
 
         # This is where is gets a little complicated. We need to do a transitive update of bazel_dep
@@ -183,19 +177,22 @@ def main():
                     new_package_version = increment_version(old_package_version)
                     sp.write("  + Incrementing {0} from {1} to {2}".format(
                         package_name, old_package_version, new_package_version))
+                    updated_modules[package_name] = new_package_version
 
                     # Copy the old module
-                    old_module_dir = args.working_directory / "modules" / package_name / old_package_version
-                    new_module_dir = args.working_directory / "modules" / package_name / new_package_version
-                    shutil.copytree(old_module_dir, new_module_dir, dirs_exist_ok=True)
+                    if not args.dry_run:
 
-                    # Update the metadata.json file
-                    metadata_json_path = args.working_directory / "modules" / package_name / "metadata.json"
-                    add_version_to_metadata_json(metadata_json_path, new_package_version)
+                        # Copy the old module
+                        old_module_dir = args.working_directory / "modules" / package_name / old_package_version
+                        new_module_dir = args.working_directory / "modules" / package_name / new_package_version
+                        shutil.copytree(old_module_dir, new_module_dir, dirs_exist_ok=True)
 
-                    # List this as an updated module.
-                    updated_modules[package_name] = new_package_version
-                    updated_at_least_one_module_version = True
+                        # Update the metadata.json file
+                        metadata_json_path = args.working_directory / "modules" / package_name / "metadata.json"
+                        add_version_to_metadata_json(metadata_json_path, new_package_version)
+
+                        # List this as an updated module.
+                        updated_at_least_one_module_version = True
 
         # Run through all MODULE.bazel files replacing any bazel_dep calls to old
         # package versions with the new package versions.
@@ -203,10 +200,11 @@ def main():
         prev_modules = {k: old_packages[k] for k in updated_modules.keys()}
         for package_name in sorted(updated_modules.keys()):
             package_new_version = updated_modules[package_name]
-            module_file = args.working_directory / "modules" / package_name / package_new_version / "MODULE.bazel"
-            update_package_dependencies(module_file, package_name, prev_modules, updated_modules)
-            sp.write("  + Updated dependencies for {0}@{1}".format(
+            sp.write("  + Updating dependencies for {0}@{1}".format(
                 package_name, package_new_version))
+            if not args.dry_run:
+                module_file = args.working_directory / "modules" / package_name / package_new_version / "MODULE.bazel"
+                update_package_dependencies(module_file, package_name, prev_modules, updated_modules)
         sp.write("> Updated {0} modules with new dependencies".format(len(updated_modules)))
 
         sp.ok("✔")
